Adapt the updated CYW43 spi bus driver to work
The previous changes did not incorporate the needed changes to the
pico cyw43 spi driver to use the slimed down SPI interface. It
fit into the Pico 2 W PIO state machine space because of it's extra
PIO but failed to fit in the original Pico W. This commit reintroduces
the code need to load and use the smaller SPI PIO program.

This is targeted for the DaynaPORT builds of the ZuluSCSI but can
be used on any RP2 build.